refactor(application): standardize use case module names
